+++ emcc.py
@@ -145,7 +145,7 @@ UBSAN_SANITIZERS = {
 }
 
 
-VALID_ENVIRONMENTS = ('web', 'webview', 'worker', 'node', 'shell')
+VALID_ENVIRONMENTS = ('web', 'webview', 'worker', 'node', 'shell', 'audioworklet')
 SIMD_INTEL_FEATURE_TOWER = ['-msse', '-msse2', '-msse3', '-mssse3', '-msse4.1', '-msse4.2', '-mavx']
 SIMD_NEON_FLAGS = ['-mfpu=neon']
 COMPILE_ONLY_FLAGS = set(['--default-obj-ext'])
@@ -294,6 +294,9 @@ def setup_environment_settings():
       not settings.ENVIRONMENT or \
       'worker' in environments or \
       (settings.ENVIRONMENT_MAY_BE_NODE and settings.USE_PTHREADS)
+
+  # Worklet environment must be enabled explicitly for now
+  settings.ENVIRONMENT_MAY_BE_AUDIOWORKLET = 'audioworklet' in environments
 
   if not settings.ENVIRONMENT_MAY_BE_WORKER and settings.PROXY_TO_WORKER:
     exit_with_error('If you specify --proxy-to-worker and specify a "-s ENVIRONMENT=" directive, it must include "worker" as a target! (Try e.g. -s ENVIRONMENT=web,worker)')
@@ -1938,6 +1941,10 @@ def phase_linker_setup(options, state, newargs, settings_map):
       'emscripten_sync_run_in_main_thread_2',
       'emscripten_sync_run_in_main_thread_4',
     ]
+    if settings.ENVIRONMENT_MAY_BE_AUDIOWORKLET:
+      settings.EXPORTED_FUNCTIONS += [
+        '_pthread_create',
+      ]
     settings.DEFAULT_LIBRARY_FUNCS_TO_INCLUDE += [
       '$exitOnMainThread',
     ]
@@ -3373,7 +3380,6 @@ var %(EXPORT_NAME)s = (function() {
     f.write(src)
 
     # Export using a UMD style export, or ES6 exports if selected
-
     if settings.EXPORT_ES6:
       f.write('export default %s;' % settings.EXPORT_NAME)
     elif not settings.MINIMAL_RUNTIME:
@@ -3385,6 +3391,12 @@ else if (typeof define === 'function' && define['amd'])
 else if (typeof exports === 'object')
   exports["%(EXPORT_NAME)s"] = %(EXPORT_NAME)s;
 ''' % {'EXPORT_NAME': settings.EXPORT_NAME})
+
+    # Store the export on the global scope for audio worklets
+    if settings.ENVIRONMENT_MAY_BE_AUDIOWORKLET:
+      f.write(f'''if (typeof AudioWorkletGlobalScope === 'function')
+  globalThis["{settings.EXPORT_NAME}"] = {settings.EXPORT_NAME};
+''')
 
   shared.configuration.get_temp_files().note(final_js)
   save_intermediate('modularized')