+++ src/library_pthread.js
@@ -224,11 +224,45 @@ var LibraryPThread = {
         PThread.tlsInitFunctions[i]();
       }
     },
-    // Loads the WebAssembly module into the given list of Workers.
+    // Builds the initial load message for the Worker
+    buildWorkerLoadMessage: function() {
+      return {
+        'cmd': 'load',
+        // If the application main .js file was loaded from a Blob, then it is not possible
+        // to access the URL of the current script that could be passed to a Web Worker so that
+        // it could load up the same file. In that case, developer must either deliver the Blob
+        // object in Module['mainScriptUrlOrBlob'], or a URL to it, so that pthread Workers can
+        // independently load up the same main application file.
+        'urlOrBlob': Module['mainScriptUrlOrBlob']
+#if !EXPORT_ES6
+        || _scriptDir
+#endif
+        ,
+#if WASM2JS
+        // the polyfill WebAssembly.Memory instance has function properties,
+        // which will fail in postMessage, so just send a custom object with the
+        // property we need, the buffer
+        'wasmMemory': { 'buffer': wasmMemory.buffer },
+#else // WASM2JS
+        'wasmMemory': wasmMemory,
+#endif // WASM2JS
+        'wasmModule': wasmModule,
+#if LOAD_SOURCE_MAP
+        'wasmSourceMap': wasmSourceMap,
+#endif
+#if USE_OFFSET_CONVERTER
+        'wasmOffsetConverter': wasmOffsetConverter,
+#endif
+#if MAIN_MODULE
+        'dynamicLibraries': Module['dynamicLibraries'],
+#endif
+      }
+    },
+    // Sets up the message handler on the passed in worker
     // onFinishedLoading: A callback function that will be called once all of
     //                    the workers have been initialized and are
     //                    ready to host pthreads.
-    loadWasmModuleToWorker: function(worker, onFinishedLoading) {
+    setupWorkerMessageHandler: function(worker, onFinishedLoading) {
       worker.onmessage = function(e) {
         var d = e['data'];
         var cmd = d['cmd'];
@@ -288,7 +322,8 @@ var LibraryPThread = {
           if (Module['onAbort']) {
             Module['onAbort'](d['arg']);
           }
-        } else {
+        }
+        else {
           err("worker sent an unknown command " + cmd);
         }
         PThread.currentProxiedOperationCallerThread = undefined;
@@ -318,39 +353,6 @@ var LibraryPThread = {
       assert(wasmMemory instanceof WebAssembly.Memory, 'WebAssembly memory should have been loaded by now!');
       assert(wasmModule instanceof WebAssembly.Module, 'WebAssembly Module should have been loaded by now!');
 #endif
-
-      // Ask the new worker to load up the Emscripten-compiled page. This is a heavy operation.
-      worker.postMessage({
-        'cmd': 'load',
-        // If the application main .js file was loaded from a Blob, then it is not possible
-        // to access the URL of the current script that could be passed to a Web Worker so that
-        // it could load up the same file. In that case, developer must either deliver the Blob
-        // object in Module['mainScriptUrlOrBlob'], or a URL to it, so that pthread Workers can
-        // independently load up the same main application file.
-        'urlOrBlob': Module['mainScriptUrlOrBlob']
-#if !EXPORT_ES6
-        || _scriptDir
-#endif
-        ,
-#if WASM2JS
-        // the polyfill WebAssembly.Memory instance has function properties,
-        // which will fail in postMessage, so just send a custom object with the
-        // property we need, the buffer
-        'wasmMemory': { 'buffer': wasmMemory.buffer },
-#else // WASM2JS
-        'wasmMemory': wasmMemory,
-#endif // WASM2JS
-        'wasmModule': wasmModule,
-#if LOAD_SOURCE_MAP
-        'wasmSourceMap': wasmSourceMap,
-#endif
-#if USE_OFFSET_CONVERTER
-        'wasmOffsetConverter': wasmOffsetConverter,
-#endif
-#if MAIN_MODULE
-        'dynamicLibraries': Module['dynamicLibraries'],
-#endif
-      });
     },
 
     // Creates a new web Worker and places it in the unused worker pool to wait for its use.
@@ -419,10 +421,115 @@ var LibraryPThread = {
 #endif
 #endif
         PThread.allocateUnusedWorker();
-        PThread.loadWasmModuleToWorker(PThread.unusedWorkers[0]);
+        PThread.setupWorkerMessageHandler(PThread.unusedWorkers[0]);
+
+        // Ask the new worker to load up the Emscripten-compiled page. This is a heavy operation.
+        PThread.unusedWorkers[0].postMessage(PThread.buildWorkerLoadMessage());
       }
       return PThread.unusedWorkers.pop();
+    },
+
+    busySpinWait: function(msecs) {
+      var t = performance.now() + msecs;
+      while (performance.now() < t) {
+        ;
+      }
+    },
+
+#if ENVIRONMENT_MAY_BE_AUDIOWORKLET
+    // Initializes a pthread in the AudioWorkletGlobalScope for this audio context
+    initAudioWorkletPThread: function(audioCtx, pthreadPtr) {
+      var aw = audioCtx.audioWorklet;
+      assert(!aw.pthread, "Can't call initAudioWorkletPThread twice on the same audio context");
+      aw.pthread = {};
+
+      // First, load the code into the worklet context. The .worker.js first, then the main script.
+#if MINIMAL_RUNTIME
+      var pthreadMainJs = Module['worker'];
+#else
+      var pthreadMainJs = locateFile('{{{ PTHREAD_WORKER_FILE }}}');
+#endif
+      aw.addModule(pthreadMainJs).then(function() {
+        // Create a dummy worklet node that we use to establish the message channel
+        // This worklet is not conected anywhere so 'process' doesn't get called so it's 
+        // not a performance overhead to have it instantiated
+        //
+        // NOTE: We pass in the 'load' message here in processorOptions because a recent
+        // Chrome change (v95+) broke `WebAssembly.Module` sending in postMessage below, but
+        // it still works if passed through `processorOptions`
+        var dummy = new AudioWorkletNode(audioCtx, 'pthread-dummy-processor', {
+          numberOfInputs: 0,
+          numberOfOutputs : 1,
+          outputChannelCount : [1],
+          processorOptions: PThread.buildWorkerLoadMessage() 
+        })
+
+        // Push this node into the PThread internal worker pool so it
+        // gets picked up in _pthread_create below. 
+        PThread.unusedWorkers.push(dummy);
+
+        // Add postMessage directly on the object, forwarded to port.postMessage (emulates Worker)
+        dummy.postMessage = dummy.port.postMessage.bind(dummy.port);
+
+        // We still call setupWorkerMessageHandler to setup the pthread environment,
+        // but we skip the actual script loading since it's already been done via
+        // addModule above.
+        PThread.setupWorkerMessageHandler(dummy);
+
+        // Forward port.onMessage to onmessage on the object (emulates Worker) but
+        // add a few worklet-only messages 
+        dummy.port.onmessage = function(e) {
+          var d = e['data'];
+          var cmd = d['cmd'];
+          if(cmd === 'addmodule') {
+            // This is sent from the worklet after the worker.js has processed the 'load' 
+            // commands and needs to load the main js (which it can't do directly in 
+            // AudioWorkletGlobalScope so we do it here from the main thread)
+            aw.addModule((Module['mainScriptUrlOrBlob'] || _scriptDir)).then(function() {
+              dummy.postMessage({'cmd': 'moduleloaded'}); 
+            });
+          } else if (cmd === 'running') {
+            // This is notified to let us know the pthread environment is ready and we can go ahead
+            // and create any pending AudioWorkletNodes
+            aw.pthread.resolveRunPromise();
+            delete aw.pthread.resolveRunPromise;
+          } else {
+            // Pass it on to the regular worker onmessage
+            dummy.onmessage(e)
+          }
+        }
+
+        // Call pthread_create to have Emscripten init our worklet pthread globals as if
+        // it was a regular worker
+        _pthread_create(pthreadPtr, 0, 0, 0);
+
+        aw.pthread.dummyWorklet = dummy;
+      }, function(err) {
+        aw.pthread.rejectRunPromise(err);
+        delete aw.pthread;
+      });
+
+      aw.pthread.runPromise = new Promise(function(resolve, reject) {
+        aw.pthread.resolveRunPromise = resolve;
+        aw.pthread.rejectRunPromise = reject;
+      });
+
+      return aw.pthread.runPromise;
+    },
+
+    // Creates an AudioWorkletNode on the specified audio context.
+    // initAudioWorkletPThread must've been called on the audio context before.
+    createAudioWorkletNode: function(audioCtx, processorName, processorOpts) {    
+      assert(audioCtx.audioWorklet.pthread.runPromise, "Call initAudioWorkletPThread once before calling createAudioWorklet");
+      return new Promise(function(resolve, reject) {
+        audioCtx.audioWorklet.pthread.runPromise.then(function() {
+          resolve(new AudioWorkletNode(audioCtx, processorName, processorOpts));
+        }, function(err) {
+          reject(err);
+        })
+      });
     }
+#endif
   },
 
   $killThread__deps: ['_emscripten_thread_free_data'],
@@ -844,8 +951,8 @@ var LibraryPThread = {
   emscripten_futex_wait__deps: ['emscripten_main_thread_process_queued_calls'],
   emscripten_futex_wait: function(addr, val, timeout) {
     if (addr <= 0 || addr > HEAP8.length || addr&3 != 0) return -{{{ cDefine('EINVAL') }}};
-    // We can do a normal blocking wait anywhere but on the main browser thread.
-    if (!ENVIRONMENT_IS_WEB) {
+    // We can do a normal blocking wait anywhere but on the main browser thread or in an audio worklet.
+    if (!ENVIRONMENT_IS_WEB && !ENVIRONMENT_IS_AUDIOWORKLET) {
 #if PTHREADS_PROFILING
       PThread.setThreadStatusConditional(_pthread_self(), {{{ cDefine('EM_THREAD_STATUS_RUNNING') }}}, {{{ cDefine('EM_THREAD_STATUS_WAITFUTEX') }}});
 #endif
@@ -882,10 +989,20 @@ var LibraryPThread = {
       // ourselves before calling the potentially-recursive call. See below for
       // how we handle the case of our futex being notified during the time in
       // between when we are not set as the value of __emscripten_main_thread_futex.
-#if ASSERTIONS
-      assert(__emscripten_main_thread_futex > 0);
+      //
+      // For audio worklets we use the same global address since they all run on 
+      // the audio thread. It's all very similar to the main thread case, except we 
+      // don't have to do any nested call special casing.
+      var theFutex = __emscripten_main_thread_futex;
+#if ENVIRONMENT_MAY_BE_AUDIOWORKLET
+      if (ENVIRONMENT_IS_AUDIOWORKLET) {
+        theFutex = __emscripten_audio_worklet_futex;
+      }
 #endif
-      var lastAddr = Atomics.exchange(HEAP32, __emscripten_main_thread_futex >> 2, addr);
+#if ASSERTIONS
+      assert(theFutex > 0);
+#endif
+      var lastAddr = Atomics.exchange(HEAP32, theFutex >> 2, addr);
 #if ASSERTIONS
       // We must not have already been waiting.
       assert(lastAddr == 0);
@@ -899,7 +1016,7 @@ var LibraryPThread = {
           PThread.setThreadStatusConditional(_pthread_self(), {{{ cDefine('EM_THREAD_STATUS_RUNNING') }}}, {{{ cDefine('EM_THREAD_STATUS_WAITFUTEX') }}});
 #endif
           // We timed out, so stop marking ourselves as waiting.
-          lastAddr = Atomics.exchange(HEAP32, __emscripten_main_thread_futex >> 2, 0);
+          lastAddr = Atomics.exchange(HEAP32, theFutex >> 2, 0);
 #if ASSERTIONS
           // The current value must have been our address which we set, or
           // in a race it was set to 0 which means another thread just allowed
@@ -908,12 +1025,25 @@ var LibraryPThread = {
 #endif
           return -{{{ cDefine('ETIMEDOUT') }}};
         }
+
+#if ENVIRONMENT_MAY_BE_AUDIOWORKLET
+        if (ENVIRONMENT_IS_AUDIOWORKLET) {
+          // Audio worklet version without any special casing like for the main thread below
+          lastAddr = Atomics.load(HEAP32, theFutex >> 2);
+          if (lastAddr != addr) {
+            // We were told to stop waiting, so stop.
+            break;
+          }
+          continue;
+        }
+#endif
+
         // We are performing a blocking loop here, so we must handle proxied
         // events from pthreads, to avoid deadlocks.
         // Note that we have to do so carefully, as we may take a lock while
         // doing so, which can recurse into this function; stop marking
         // ourselves as waiting while we do so.
-        lastAddr = Atomics.exchange(HEAP32, __emscripten_main_thread_futex >> 2, 0);
+        lastAddr = Atomics.exchange(HEAP32, theFutex >> 2, 0);
 #if ASSERTIONS
         assert(lastAddr == addr || lastAddr == 0);
 #endif
@@ -983,33 +1113,46 @@ var LibraryPThread = {
     // For Atomics.notify() API Infinity is to be passed in that case.
     if (count >= {{{ cDefine('INT_MAX') }}}) count = Infinity;
 
-    // See if main thread is waiting on this address? If so, wake it up by resetting its wake location to zero.
-    // Note that this is not a fair procedure, since we always wake main thread first before any workers, so
+    // See if any spinning thread is waiting on this address? If so, wake it up by resetting its wake location to zero.
+    // Note that this is not a fair procedure, since we always wake these threads first before any workers, so
     // this scheme does not adhere to real queue-based waiting.
-#if ASSERTIONS
-    assert(__emscripten_main_thread_futex > 0);
+    // Spin-wait futexes are used on the main thread and in worklets due to the lack of Atomic.wait().
+    var spinFutexesWoken = 0;
+    
+#if ENVIRONMENT_MAY_BE_AUDIOWORKLET
+    var spinFutexes = [__emscripten_main_thread_futex, __emscripten_audio_worklet_futex];
+    for (var i = 0; i < spinFutexes.length; i++) {
+      var theFutex = spinFutexes[i];
+#else
+    var theFutex = __emscripten_main_thread_futex;
 #endif
-    var mainThreadWaitAddress = Atomics.load(HEAP32, __emscripten_main_thread_futex >> 2);
-    var mainThreadWoken = 0;
-    if (mainThreadWaitAddress == addr) {
+#if ASSERTIONS
+    assert(theFutex > 0);
+#endif
+    var waitAddress = Atomics.load(HEAP32, theFutex >> 2);
+    
+    if (waitAddress == addr) {
 #if ASSERTIONS
       // We only use __emscripten_main_thread_futex on the main browser thread, where we
       // cannot block while we wait. Therefore we should only see it set from
       // other threads, and not on the main thread itself. In other words, the
       // main thread must never try to wake itself up!
-      assert(!ENVIRONMENT_IS_WEB);
+      assert(theFutex != __emscripten_main_thread_futex || !ENVIRONMENT_IS_WEB);
 #endif
-      var loadedAddr = Atomics.compareExchange(HEAP32, __emscripten_main_thread_futex >> 2, mainThreadWaitAddress, 0);
-      if (loadedAddr == mainThreadWaitAddress) {
+      var loadedAddr = Atomics.compareExchange(HEAP32, theFutex >> 2, waitAddress, 0);
+      if (loadedAddr == waitAddress) {
         --count;
-        mainThreadWoken = 1;
+        spinFutexesWoken += 1;
         if (count <= 0) return 1;
       }
     }
+#if ENVIRONMENT_MAY_BE_AUDIOWORKLET
+    } // close out the spinFutexes loop
+#endif
 
     // Wake any workers waiting on this address.
     var ret = Atomics.notify(HEAP32, addr >> 2, count);
-    if (ret >= 0) return ret + mainThreadWoken;
+    if (ret >= 0) return ret + spinFutexesWoken;
     throw 'Atomics.notify returned an unexpected value ' + ret;
   },