+++ src/worker.js
@@ -50,6 +50,42 @@ if (typeof process === 'object' && typeof process.versions === 'object' && typeo
 }
 #endif // ENVIRONMENT_MAY_BE_NODE
 
+#if ENVIRONMENT_MAY_BE_AUDIOWORKLET
+if (typeof AudioWorkletGlobalScope === "function") {
+  Object.assign(globalThis, {
+    Module: Module,
+    self: globalThis, // Unlike DedicatedWorkerGlobalScope, AudioWorkletGlobalScope doesn't have 'self'
+    performance: { // Polyfill performance.now() since it's missing in worklets, falling back to Date.now()
+      start: Date.now(),
+      now: function() {
+        return Date.now() - performance.start;
+      }
+    }
+  });
+    
+  // Create a dummy processor which we use to bootstrap PThreads in the worklet context
+  class PThreadDummyProcessor extends AudioWorkletProcessor {
+    constructor(arg) {
+      super();
+      // Make message passing work directly on the global scope to simulate what regular
+      // Workers do which makes all the code above work the same.
+      this.port.onmessage = self.onmessage;
+      self.postMessage = this.port.postMessage.bind(this.port);
+
+      // Pass the initial message to the message handler as if it was sent via `sendMessage`
+      self.onmessage({
+        data: arg.processorOptions
+      });
+    }
+    
+    // Needs a dummy process method too otherwise 'registerProcessor' fails below
+    process(inputs, outputs, parameters) {}
+  }
+  
+  registerProcessor('pthread-dummy-processor', PThreadDummyProcessor);
+}
+#endif
+
 // Thread-local:
 #if EMBIND
 var initializedJS = false; // Guard variable for one-time init of the JS state (currently only embind types registration)
@@ -134,6 +170,17 @@ self.onmessage = function(e) {
 
 #if !MINIMAL_RUNTIME || MODULARIZE
       {{{ makeAsmImportsAccessInPthread('ENVIRONMENT_IS_PTHREAD') }}} = true;
+#if ENVIRONMENT_MAY_BE_AUDIOWORKLET
+      {{{ makeAsmImportsAccessInPthread('ENVIRONMENT_IS_AUDIOWORKLET') }}} = typeof AudioWorkletGlobalScope === "function";
+#endif
+#endif
+
+#if ENVIRONMENT_MAY_BE_AUDIOWORKLET
+      // When running as an AudioWorklet all the scripts are imported from the main thread (via .addModule)      
+      if({{{ makeAsmImportsAccessInPthread('ENVIRONMENT_IS_AUDIOWORKLET') }}}) { 
+        postMessage({'cmd': 'addmodule'});
+        return
+      }
 #endif
 
 #if MODULARIZE && EXPORT_ES6
@@ -174,6 +221,20 @@ self.onmessage = function(e) {
 #endif
 #endif
 #endif // MODULARIZE && EXPORT_ES6
+#if ENVIRONMENT_MAY_BE_AUDIOWORKLET
+    } else if (e.data.cmd === 'moduleloaded') {
+      #if MODULARIZE
+      #if MINIMAL_RUNTIME
+            {{{ EXPORT_NAME }}}(imports).then(function (instance) {
+              Module = instance;
+            });
+      #else
+            {{{ EXPORT_NAME }}}(Module).then(function (instance) {
+              Module = instance;
+            });
+      #endif
+      #endif
+#endif // ENVIRONMENT_MAY_BE_AUDIOWORKLET
     } else if (e.data.cmd === 'run') {
       // This worker was idle, and now should start executing its pthread entry
       // point.
@@ -207,6 +268,12 @@ self.onmessage = function(e) {
       }
 #endif // EMBIND
 
+#if ENVIRONMENT_MAY_BE_AUDIOWORKLET
+      if (Module['ENVIRONMENT_IS_AUDIOWORKLET']) {
+        // Audio worklets don't run any entrypoint since their entry points are the 'process' function invocations
+        postMessage({'cmd': 'running'});
+      } else {
+#endif
       try {
         // pthread entry points are always of signature 'void *ThreadMain(void *arg)'
         // Native codebases sometimes spawn threads with other thread entry point signatures,
@@ -276,6 +343,9 @@ self.onmessage = function(e) {
 #endif
         }
       }
+#if ENVIRONMENT_MAY_BE_AUDIOWORKLET
+      }
+#endif
     } else if (e.data.cmd === 'cancel') { // Main thread is asking for a pthread_cancel() on this thread.
       if (Module['_pthread_self']()) {
         Module['__emscripten_thread_exit'](-1/*PTHREAD_CANCELED*/);